requirements.txt
alabaster==0.7.10
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.22.0
attrs==17.1.0
Automat==0.6.0
autopep8==1.3.1
Babel==2.4.0
backports-abc==0.5
beautifulsoup4==4.6.0
bleach==2.0.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
cycler==0.10.0
decorator==4.0.11
docutils==0.13.1
entrypoints==0.2.2
future==0.16.0
html5lib==0.999999999
idna==2.5
imagesize==0.7.1
incremental==16.10.1
ipykernel==4.6.0
ipyparallel==6.0.2
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lxml==3.7.3
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
nbconvert==5.1.1
nbformat==4.3.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numpy==1.12.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
parsel==1.2.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
plotly==2.0.8
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pycparser==2.17
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-twitter==3.3
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
queuelib==1.4.2
requests==2.16.5
requests-oauthlib==0.8.0
scipy==0.19.0
Scrapy==1.3.3
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.5
sympy==1.0
terminado==0.6
testpath==0.3
tornado==4.4.3
traitlets==4.3.2
Twisted==17.1.0
typing==3.6.1
urllib3==1.21.1
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
zope.interface==4.4.1